app/requirements.txt
discord.py
yt-dlp
mutagen